// frontend/sdk/models/PlaylistDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sopranino API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PlaylistDto
 */
export interface PlaylistDto {
    /**
     * The ID of the playlist
     * @type {string}
     * @memberof PlaylistDto
     */
    id: string;
    /**
     * The name of the playlist
     * @type {string}
     * @memberof PlaylistDto
     */
    name: string;
    /**
     * The description of the playlist
     * @type {string}
     * @memberof PlaylistDto
     */
    description?: string | null;
    /**
     * The image URL of the playlist
     * @type {string}
     * @memberof PlaylistDto
     */
    imageUrl: string;
    /**
     * The owner of the playlist
     * @type {string}
     * @memberof PlaylistDto
     */
    owner: string;
    /**
     * The total number of tracks in the playlist
     * @type {number}
     * @memberof PlaylistDto
     */
    totalTracks: number;
    /**
     * Whether the playlist is public
     * @type {boolean}
     * @memberof PlaylistDto
     */
    isPublic: boolean;
    /**
     * The external URL of the playlist
     * @type {string}
     * @memberof PlaylistDto
     */
    externalUrl: string;
}

/**
 * Check if a given object implements the PlaylistDto interface.
 */
export function instanceOfPlaylistDto(value: object): value is PlaylistDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('imageUrl' in value) || value['imageUrl'] === undefined) return false;
    if (!('owner' in value) || value['owner'] === undefined) return false;
    if (!('totalTracks' in value) || value['totalTracks'] === undefined) return false;
    if (!('isPublic' in value) || value['isPublic'] === undefined) return false;
    if (!('externalUrl' in value) || value['externalUrl'] === undefined) return false;
    return true;
}

export function PlaylistDtoFromJSON(json: any): PlaylistDto {
    return PlaylistDtoFromJSONTyped(json, false);
}

export function PlaylistDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlaylistDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'imageUrl': json['imageUrl'],
        'owner': json['owner'],
        'totalTracks': json['totalTracks'],
        'isPublic': json['isPublic'],
        'externalUrl': json['externalUrl'],
    };
}

export function PlaylistDtoToJSON(json: any): PlaylistDto {
    return PlaylistDtoToJSONTyped(json, false);
}

export function PlaylistDtoToJSONTyped(value?: PlaylistDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
        'imageUrl': value['imageUrl'],
        'owner': value['owner'],
        'totalTracks': value['totalTracks'],
        'isPublic': value['isPublic'],
        'externalUrl': value['externalUrl'],
    };
}